third_party/WebKit/LayoutTests/fast/dom/shadow/istabstop-property.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<p>This is basic behavior test for isTabStop attribute.</p>
<pre id="console"></pre>
<script>
var div;

function test() {
    debug("Testing isTabStop property and attribute");

    debug("Test isTabStop normal assignment behavior");
    div = document.createElement("div");
    var flag = div.isTabStop;
    shouldBe("div.isTabStop", "false");

    div.isTabStop = flag;
    shouldBe("div.isTabStop", "false");

    div.removeAttribute('isTabstop');
    shouldBe("div.isTabStop", "false");

    debug("Test isTabStop override by tabindex property");
    div.tabIndex = 0;
    shouldBe("div.isTabStop", "true");

    div.tabIndex = -1;
    shouldBe("div.isTabStop", "false");

    div.tabIndex = 1;
    shouldBe("div.isTabStop", "true");

    debug("Test isTabStop override by tabindex attribute");
    div.isTabStop = false;
    div.setAttribute("tabindex", "0");
    shouldBe("div.isTabStop", "true");

    div.isTabStop = true;
    div.setAttribute("tabindex", "-1");
    shouldBe("div.isTabStop", "false");

    div.isTabStop = false;
    div.setAttribute("tabIndex", "1");
    shouldBe("div.isTabStop", "true");

    debug("Test isTabStop change after tabindex change");
    div.tabIndex = 0;
    div.isTabStop = false;
    shouldBe("div.isTabStop", "false");

    div.tabIndex = -1;
    div.isTabStop = true;
    shouldBe("div.isTabStop", "true");

    debug("Test isTabStop change before tabindex change which will be overridden");
    div.isTabStop = false;
    div.tabIndex = 0;
    shouldBe("div.isTabStop", "true");

    div.isTabStop = true;
    div.tabIndex = -1;
    shouldBe("div.isTabStop", "false");
}

function run_test() {
    if (window.testRunner)
        testRunner.dumpAsText();

    test();

    debug('Test finished.');
}

run_test();
</script>
</body>
</html>
