third_party/WebKit/LayoutTests/fast/dom/shadow/istabstop-property-expected.txt

This is basic behavior test for isTabStop attribute.

Testing isTabStop property and attribute
Test isTabStop normal assignment behavior
PASS div.isTabStop is false
PASS div.isTabStop is false
PASS div.isTabStop is false
Test isTabStop override by tabindex property
PASS div.isTabStop is true
PASS div.isTabStop is false
PASS div.isTabStop is true
Test isTabStop override by tabindex attribute
PASS div.isTabStop is true
PASS div.isTabStop is false
PASS div.isTabStop is true
Test isTabStop change after tabindex change
PASS div.isTabStop is false
PASS div.isTabStop is true
Test isTabStop change before tabindex change which will be overridden
PASS div.isTabStop is true
PASS div.isTabStop is false
Test finished.
PASS successfullyParsed is true

TEST COMPLETE
